requirements.txt
joblib
nltk
scikit-learn==1.6.1
duckduckgo-search
google-generativeai